extract draw node to own file
